requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
Pillow==10.0.1
drf-yasg==1.21.5
python-decouple==3.8
django-cleanup==8.0.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
Faker==19.6.2